+++ demo/src/main/java/com/SistemaVenta/demo/Repositorios/IDetailsSaleRepository.java
@@ -2,6 +2,6 @@ package com.SistemaVenta.demo.Repositorios;
 import com.SistemaVenta.demo.Model.DetailsSale;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IDetailsSaleRepository extends JpaRepository <DetailsSale,Long>{
+public interface IDetailsSaleRepository extends JpaRepository <DetailsSale,Integer>{
 
 }
