+++ demo/src/main/java/com/SistemaVenta/demo/Repositorios/IRoleRepository.java
@@ -4,7 +4,7 @@ package com.SistemaVenta.demo.Repositorios;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.SistemaVenta.demo.Model.Role;
 
-public interface IRoleRepository extends JpaRepository<Role, Long> {
+public interface IRoleRepository extends JpaRepository<Role, Integer> {
     
 
 }
